Repository paseos

+++ TP2_Grupo11/src/main/java/ar/edu/unju/fi/Listas/ListaHorario.java
@@ -15,12 +15,12 @@ public class ListaHorario{
     public ListaHorario(){
         horarios = new ArrayList<Turno>();
         //precarga de horarios
-        horarios.add(new Turno(1,"Lunes","9 a 11 hrs","15 a 19 hrs","juan Perez","Maria Juana"));
-        horarios.add(new Turno(2,"Martes","9 a 11 hrs","15 a 19 hrs","juan Perez","Maria Juana"));
-        horarios.add(new Turno(3,"Miércoles","9 a 11 hrs","15 a 19 hrs","juan Perez","Roberto Vasquez"));
-        horarios.add(new Turno(4,"Jueves","9 a 11 hrs","15 a 19 hrs","Lucas Días","Maria Juana"));
-        horarios.add(new Turno(5,"Viernes","9 a 11 hrs","15 a 19 hrs","juan Perez","Maria Juana"));
-        horarios.add(new Turno(6,"Sábado","9 a 11 hrs","15 a 19 hrs","johan Rosero","Maria Juana"));
+        horarios.add(new Turno((long) 1,"Lunes","9 a 11 hrs","15 a 19 hrs","juan Perez","Maria Juana"));
+        horarios.add(new Turno((long) 2,"Martes","9 a 11 hrs","15 a 19 hrs","juan Perez","Maria Juana"));
+        horarios.add(new Turno((long) 3,"Miércoles","9 a 11 hrs","15 a 19 hrs","juan Perez","Roberto Vasquez"));
+        horarios.add(new Turno((long) 4,"Jueves","9 a 11 hrs","15 a 19 hrs","Lucas Días","Maria Juana"));
+        horarios.add(new Turno((long) 5,"Viernes","9 a 11 hrs","15 a 19 hrs","juan Perez","Maria Juana"));
+        horarios.add(new Turno((long) 6,"Sábado","9 a 11 hrs","15 a 19 hrs","johan Rosero","Maria Juana"));
 
     }
